Add script to install hooks for Git
`run-hook` takes care of running all hooks of a certain type. The hooks
must be located in a directory inside the `githook/` directory.

`setup-hooks` creates links for all different hooks to `run-hook`.
`run-hook` can then infer the type of hook to run by retrieving the
filename of the symbolic link.